# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Barber.uz CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List barber's services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{serviceId}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /settings/profile:
    get:
      operationId: getProfile
      tags: [settings]
      summary: Get barber profile settings
      responses:
        "200":
          description: Profile settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [settings]
      summary: Update profile settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /settings/password:
    put:
      operationId: updatePassword
      tags: [settings]
      summary: Update password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Wrong old password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings/notifications:
    get:
      operationId: getNotificationSettings
      tags: [settings]
      summary: Get notification settings
      responses:
        "200":
          description: Notification settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    put:
      operationId: updateNotificationSettings
      tags: [settings]
      summary: Update notification settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        barberId:
          type: string
          format: uuid
        name:
          type: string
        nameRu:
          type: string
          nullable: true
        duration:
          type: integer
          description: Duration in minutes
        price:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, barberId, name, duration, price, isActive, createdAt]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        nameRu:
          type: string
          nullable: true
        duration:
          type: integer
          minimum: 5
        price:
          type: number
          minimum: 0
      required: [name, duration, price]
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        nameRu:
          type: string
          nullable: true
        duration:
          type: integer
        price:
          type: number
        isActive:
          type: boolean
    ServiceListResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
      required: [services]
...
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        username:
          type: string
        brandName:
          type: string
          nullable: true
        mode:
          type: string
          enum: [solo, team]
        lang:
          type: string
          enum: [uz, ru]
        workingHoursStart:
          type: string
          nullable: true
        workingHoursEnd:
          type: string
          nullable: true
        telegramVerified:
          type: boolean
        telegramUsername:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [id, name, username, mode, lang, telegramVerified]
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        brandName:
          type: string
          nullable: true
        mode:
          type: string
          enum: [solo, team]
        lang:
          type: string
          enum: [uz, ru]
        workingHoursStart:
          type: string
          nullable: true
        workingHoursEnd:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
    UpdatePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [oldPassword, newPassword]
    NotificationSettings:
      type: object
      properties:
        newBooking:
          type: boolean
        cancellation:
          type: boolean
        reminders:
          type: boolean
        reminderMinutes:
          type: integer
          default: 30
      required: [newBooking, cancellation, reminders]