# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber.uz CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  barberId: string;
  name: string;
  nameRu?: string | null;
  /** Duration in minutes */
  duration: number;
  price: number;
  isActive: boolean;
  createdAt: Date;
}